requirements.txt
pandas==1.5.3
numpy==1.21.6
scikit-learn==1.0.2
tensorflow==2.11.0
xgboost==1.7.4
matplotlib==3.5.3
seaborn==0.12.2
plotly==5.13.1
flask==2.3.2
streamlit==1.22.0
pyyaml==6.0
joblib==1.2.0